Move EmptyStringLiteralInspectionTests for QuickFixes into relevant files under RD.QuickFixes
