+++ RubberduckTests/QuickFixes/ReplaceEmptyStringLiteralStatementQuickFixTests.cs
@@ -1,0 +1,43 @@
+﻿using System.Linq;
+using Microsoft.VisualStudio.TestTools.UnitTesting;
+using RubberduckTests.Mocks;
+using System.Threading;
+using Rubberduck.Inspections.Concrete;
+using Rubberduck.Inspections.QuickFixes;
+using Rubberduck.VBEditor.SafeComWrappers.Abstract;
+using RubberduckTests.Inspections;
+
+namespace RubberduckTests.QuickFixes
+{
+    [TestClass]
+    public class ReplaceEmptyStringLiteralStatementQuickFixTests
+    {
+        [TestMethod]
+        [TestCategory("Inspections")]
+        public void EmptyStringLiteral_QuickFixWorks()
+        {
+            const string inputCode =
+@"Public Sub Foo(ByRef arg1 As String)
+    arg1 = """"
+End Sub";
+
+            const string expectedCode =
+@"Public Sub Foo(ByRef arg1 As String)
+    arg1 = vbNullString
+End Sub";
+
+            IVBComponent component;
+            var vbe = MockVbeBuilder.BuildFromSingleStandardModule(inputCode, out component);
+            var state = MockParser.CreateAndParse(vbe.Object);
+
+            var inspection = new EmptyStringLiteralInspection(state);
+            var inspector = InspectionsHelper.GetInspector(inspection);
+            var inspectionResults = inspector.FindIssuesAsync(state, CancellationToken.None).Result;
+
+            new ReplaceEmptyStringLiteralStatementQuickFix(state).Fix(inspectionResults.First());
+
+            Assert.AreEqual(expectedCode, state.GetRewriter(component).GetText());
+        }
+
+    }
+}
